Improve documentation for the crypto module.
