+++ src/crypto/ring.rs
@@ -99,9 +99,11 @@ impl std::error::Error for GenerateError {}
 
 //----------- DigestContext --------------------------------------------------
 
+/// Context for computing a message digest.
 pub struct DigestContext(Context);
 
 impl DigestContext {
+    /// Create a new context for a specified digest type.
     pub fn new(digest_type: DigestType) -> Self {
         Self(match digest_type {
             DigestType::Sha1 => Context::new(&SHA1_FOR_LEGACY_USE_ONLY),
@@ -110,10 +112,12 @@ impl DigestContext {
         })
     }
 
+    /// Add input to the digest computation.
     pub fn update(&mut self, data: &[u8]) {
         self.0.update(data)
     }
 
+    /// Finish computing the digest.
     pub fn finish(self) -> Digest {
         Digest(self.0.finish())
     }
@@ -121,6 +125,7 @@ impl DigestContext {
 
 //----------- Digest ---------------------------------------------------------
 
+/// A message digest.
 pub struct Digest(RingDigest);
 
 impl AsRef<[u8]> for Digest {
@@ -131,8 +136,12 @@ impl AsRef<[u8]> for Digest {
 
 //----------- PublicKey ------------------------------------------------------
 
+/// A public key for verifying a signature.
 pub enum PublicKey {
+    /// Variant for RSA public keys.
     Rsa(&'static RsaParameters, PublicKeyComponents<Vec<u8>>),
+
+    /// Variant for elliptic-curve public keys.
     Unparsed(
         &'static dyn VerificationAlgorithm,
         UnparsedPublicKey<Vec<u8>>,
@@ -140,6 +149,7 @@ pub enum PublicKey {
 }
 
 impl PublicKey {
+    /// Create a public key from a [`Dnskey`].
     pub fn from_dnskey(
         dnskey: &Dnskey<impl AsRef<[u8]>>,
     ) -> Result<Self, AlgorithmError> {
@@ -205,6 +215,7 @@ impl PublicKey {
         }
     }
 
+    /// Verify a signature.
     pub fn verify(
         &self,
         signed_data: &[u8],
@@ -222,6 +233,7 @@ impl PublicKey {
         .map_err(|_| AlgorithmError::BadSig)
     }
 
+    /// Convert to a [`Dnskey`].
     pub fn dnskey(&self, flags: u16) -> Dnskey<Vec<u8>> {
         match self {
             PublicKey::Rsa(parameters, components) => {
@@ -327,7 +339,8 @@ impl PublicKey {
 }
 
 #[cfg(feature = "unstable-crypto-sign")]
-pub(crate) mod sign {
+/// Submodule for private keys and signing.
+pub mod sign {
     use std::boxed::Box;
     use std::sync::Arc;
     use std::vec::Vec;
@@ -356,22 +369,37 @@ pub(crate) mod sign {
     pub enum KeyPair {
         /// An RSA/SHA-256 keypair.
         RsaSha256 {
+            /// They RSA key.
             key: RsaKeyPair,
+
+            /// Flags from [`Dnskey`].
             flags: u16,
+
+            /// Random number generator.
             rng: Arc<dyn ring::rand::SecureRandom>,
         },
 
         /// An ECDSA P-256/SHA-256 keypair.
         EcdsaP256Sha256 {
+            /// The ECDSA key.
             key: EcdsaKeyPair,
+
+            /// Flags from [`Dnskey`].
             flags: u16,
+
+            /// Random number generator.
             rng: Arc<dyn ring::rand::SecureRandom>,
         },
 
         /// An ECDSA P-384/SHA-384 keypair.
         EcdsaP384Sha384 {
+            /// The ECDSA key.
             key: EcdsaKeyPair,
+
+            /// Flags from [`Dnskey`].
             flags: u16,
+
+            /// Random number generator.
             rng: Arc<dyn ring::rand::SecureRandom>,
         },
